Self_RAG/requirements.txt
langchain
langchain-community
langchainhub
streamlit
faiss-cpu
groq
python-dotenv
langchain-groq
